function/requirements.txt
azure-functions
psycopg2
psycopg2-binary
python-http-client
sendgrid
starkbank-ecdsa
